requirements.txt
xrpl-py
xrpl-dex-sdk
requests
websocket-client
python-telegram-bot
